updated weights a little, started work on number hotkeys, fixed mouse sens option, changed recoil, and removed all your couts

+++ src/keys.hpp
@@ -13,7 +13,7 @@ class Keys{
 public: //will do enum later
 	Keys();
 	//~Keys();
-	int mouseSens;
+	float mouseSens;
 	map <std::string, pair<ButtonHandle, bool> > keybinds; //world.move
 	map <std::string, void(*)(const Event* , void*)> wildKeys; //window -> get_panda_framework() -> define_key
